+++ test/Dommel.IntegrationTests/Dommel.IntegrationTests.csproj
@@ -6,6 +6,7 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="mysqlconnector" Version="0.47.1" />
+    <PackageReference Include="Npgsql" Version="4.0.3" />
     <ProjectReference Include="..\..\src\Dommel\Dommel.csproj" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.9.0" />
     <PackageReference Include="xunit" Version="2.4.0" />
